refactor: service 구조 수정
기존 Read / Write 로 분리하던 서비스를 하나의 서비스 클래스로 통합하였음(관리 용이)
