+++ src/main/java/com/antk7894/amatda/controller/DailyController.java
@@ -5,9 +5,8 @@ import com.antk7894.amatda.dto.daily.DailyPatchRequestDto;
 import com.antk7894.amatda.dto.daily.DailyUpdateRequestDto;
 import com.antk7894.amatda.entity.Daily;
 import com.antk7894.amatda.entity.planner.Planner;
-import com.antk7894.amatda.service.DailyReadService;
 import com.antk7894.amatda.service.DailyService;
-import com.antk7894.amatda.service.PlannerReadService;
+import com.antk7894.amatda.service.PlannerService;
 import com.antk7894.amatda.util.CustomSecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -25,24 +24,23 @@ import java.net.URI;
 public class DailyController {
 
     private final DailyService dailyService;
-    private final DailyReadService dailyReadService;
-    private final PlannerReadService plannerReadService;
+    private final PlannerService plannerService;
 
     @GetMapping
     public ResponseEntity<Page<Daily>> all(Pageable pageable) {
-        Planner planner = plannerReadService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
-        return ResponseEntity.ok(dailyReadService.findAll(planner, pageable));
+        Planner planner = plannerService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
+        return ResponseEntity.ok(dailyService.findAll(planner, pageable));
     }
 
     @GetMapping("/{dailyId}")
     public ResponseEntity<Daily> one(@PathVariable Long dailyId) {
-        Planner planner = plannerReadService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
-        return ResponseEntity.ok(dailyReadService.findOneById(planner, dailyId));
+        Planner planner = plannerService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
+        return ResponseEntity.ok(dailyService.findOneById(planner, dailyId));
     }
 
     @PostMapping
     public ResponseEntity<Daily> newDaily(@RequestBody DailyCreateRequestDto dto) {
-        Planner planner = plannerReadService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
+        Planner planner = plannerService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
         Daily daily = dailyService.saveOne(planner, dto);
         String currentUri = ServletUriComponentsBuilder.fromCurrentRequestUri().toUriString();
         return ResponseEntity.created(URI.create(currentUri + "/" + daily.getDailyId())).body(daily);
@@ -50,13 +48,13 @@ public class DailyController {
 
     @PutMapping("/{dailyId}")
     public ResponseEntity<Daily> updateDaily(@PathVariable Long dailyId, @RequestBody DailyUpdateRequestDto dto) {
-        Planner planner = plannerReadService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
+        Planner planner = plannerService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
         return ResponseEntity.ok(dailyService.updateOne(planner, dailyId, dto));
     }
 
     @PatchMapping("/{dailyId}")
     public ResponseEntity<Daily> manipulateDaily(@PathVariable Long dailyId, @RequestBody DailyPatchRequestDto dto) {
-        Planner planner = plannerReadService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
+        Planner planner = plannerService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
         return switch (dto.action()) {
             case FINISH -> ResponseEntity.ok(dailyService.finishOne(planner, dailyId));
             case RESET -> ResponseEntity.ok(dailyService.resetOne(planner, dailyId));
@@ -65,7 +63,7 @@ public class DailyController {
 
     @DeleteMapping("/{dailyId}")
     public ResponseEntity<?> deleteDaily(@PathVariable Long dailyId) {
-        Planner planner = plannerReadService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
+        Planner planner = plannerService.findOneByEmail(CustomSecurityUtil.getCurrentUserEmail());
         dailyService.removeOne(planner, dailyId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
     }
