+++ src/main/resources/application.yml
@@ -7,7 +7,7 @@ spring:
   jpa:
     open-in-view: false
     hibernate:
-      ddl-auto: ${DDL_AUTH}
+      ddl-auto: create
     properties:
       hibernate:
         default_batch_fetch_size: 100
